migrating streams spec to ts
