+++ packages/server/modules/core/services/streams.js
@@ -22,7 +22,7 @@ const { omitBy, isNull, isUndefined, has } = require('lodash')
 module.exports = {
   /**
    * @param {import('@/modules/core/graph/generated/graphql').StreamCreateInput & {ownerId: string}} param0
-   * @returns
+   * @returns {Promise<string>}
    */
   async createStream({
     name,
